backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.110.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lazy-model==0.2.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.5.0
pydantic-core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyMySQL==1.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
openpyxl==3.1.5
scikit-learn==1.5.2
gensim==4.3.0
scipy==1.12.0
# Rasa NLU dependencies (optional - for Rasa Intent Classifier)
# Note: Rasa is optional. If not installed, the system will use an enhanced fallback classifier
# rasa>=3.6.0
# tensorflow>=2.13.0,<2.14.0
# tensorflow-text>=2.13.0,<2.14.0
# tensorflow-addons>=0.20.0
# sanic>=23.3.0
# cloudpickle>=2.0.0
# coloredlogs>=15.0

# Additional dependencies for Rasa fallback classifier (required)
pyyaml>=6.0
numpy>=1.23.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.37.8
starlette==0.36.3
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.29.0
wrapt==1.17.2
rapidFuzz>=3.0.0
langgraph==1.1.10
langgraph-sdk==0.3.13
redis==5.2.1
httpx>=0.25.0
